// src/app/precisa/mensajeria/formato-mensaje/model/Dtomensaje.ts
export class Dtomensaje {
    constructor() {

    }

    IdMensaje?: number;  
    Codigo: string;
    Nombre: string;
    IdEmpresa?: number;
    TipoMensaje?: number; 
    Asunto: string;
    Mensaje: string;
    UsuarioCreacion: string;
    Estado?: number;  
    UsuarioModificacion: string;
    FechaCreacion: Date;
    FechaModificacion: Date;
    IpCreacion: string;
    IpModificacion: string; 
    
    NombreCompleto: string;
    CompaniaCodigo: string;
    DescTipoMensaje: string;
    DesEstado: string;

    // CODIGOS DE ANGULAR
    num: number;
  }